add optimization_util files
